src/knowledge/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
azure-common==1.1.28
azure-core==1.38.2
azure-identity==1.25.2
azure-search-documents==11.6.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
distro==1.9.0
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
jiter==0.13.0
msal==1.35.0
msal-extensions==1.3.1
openai==2.24.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
